+++ .gitignore
@@ -1,1 +1,4 @@
-
+efBike/Migrations/
+efBike/bin/
+efBike/obj/Debug/
+.vs/
