+++ efBike/Program.cs
@@ -1,2 +1,20 @@
-﻿// See https://aka.ms/new-console-template for more information
+﻿// using efBike.Classes;
+// using Microsoft.EntityFrameworkCore;
+
+// var builder = WebApplication.CreateBuilder(args);
+
+// // Add services to the container.
+
+// builder.Services.AddDbContext<BikeShopContext>(options =>
+//     options.UseSqlServer(builder.Configuration.GetConnectionString("BikeShopDatabase")));
+
+// // Other service configurations...
+
+// var app = builder.Build();
+
+// // Configure the HTTP request pipeline...
+
+// app.Run();
+
+
 Console.WriteLine("Hello, World!");
