+++ efBike/appsettings.json
@@ -1,0 +1,5 @@
+{
+  "ConnectionStrings": {
+    "BikeRepairDatabase": "Server=(localdb)\\mssqllocaldb;Database=BikeRepairDb;Trusted_Connection=True;"
+  }
+}
