--- efBike/classes/Customer.cs
@@ -1,10 +1,0 @@
-namespace efBike.Classes;
-
-public class Customer {
-  public int Id { get; set; }
-  public string Name { get; set; }
-  public string PhoneNumber { get; set; }
-  public string Email { get; set; }
-
-  public List<Order> Orders { get; set; }
-}
